requirements.txt
wmi>=1.5.1
pywin32>=306
pystray>=0.19.5
Pillow>=10.0.0
pytest>=7.4.0
